Remove unused config.  * identity  * timer (only use QuartzTimer)  * isBootnode
