+++ services/serviceutils/src/main/java/tech/pegasys/artemis/service/serviceutils/ServiceConfig.java
@@ -16,21 +16,14 @@ package tech.pegasys.artemis.service.serviceutils;
 import com.google.common.eventbus.EventBus;
 import io.vertx.core.Vertx;
 import java.util.Objects;
-import org.apache.tuweni.crypto.SECP256K1;
 import tech.pegasys.artemis.util.config.ArtemisConfiguration;
 import tech.pegasys.pantheon.metrics.MetricsSystem;
-import tech.pegasys.pantheon.metrics.noop.NoOpMetricsSystem;
 
 public class ServiceConfig {
   Vertx vertx;
   EventBus eventBus;
   MetricsSystem metricsSystem;
   ArtemisConfiguration config;
-  SECP256K1.KeyPair keyPair;
-
-  public ServiceConfig() {
-    this.metricsSystem = new NoOpMetricsSystem();
-  }
 
   public ServiceConfig(
       EventBus eventBus, Vertx vertx, MetricsSystem metricsSystem, ArtemisConfiguration config) {
@@ -38,7 +31,6 @@ public class ServiceConfig {
     this.vertx = vertx;
     this.metricsSystem = metricsSystem;
     this.config = config;
-    this.keyPair = config.getKeyPair();
   }
 
   public EventBus getEventBus() {
@@ -61,10 +53,6 @@ public class ServiceConfig {
     this.config = config;
   }
 
-  public SECP256K1.KeyPair getKeyPair() {
-    return this.keyPair;
-  }
-
   public MetricsSystem getMetricsSystem() {
     return metricsSystem;
   }
@@ -77,12 +65,11 @@ public class ServiceConfig {
     }
     ServiceConfig serviceConfig = (ServiceConfig) o;
     return Objects.equals(eventBus, serviceConfig.eventBus)
-        && Objects.equals(config, serviceConfig.config)
-        && Objects.equals(keyPair, serviceConfig.keyPair);
+        && Objects.equals(config, serviceConfig.config);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(eventBus, config, keyPair);
+    return Objects.hash(eventBus, config);
   }
 }
