+++ util/src/main/java/tech/pegasys/artemis/util/time/Timer.java
@@ -13,25 +13,69 @@
 
 package tech.pegasys.artemis.util.time;
 
-@SuppressWarnings({"rawtypes"})
-public interface Timer {
+import static org.quartz.JobBuilder.newJob;
+import static org.quartz.SimpleScheduleBuilder.simpleSchedule;
+import static org.quartz.TriggerBuilder.newTrigger;
 
-  String QUARTZ_TIMER_NAME = "QuartzTimer";
+import com.google.common.eventbus.EventBus;
+import java.util.UUID;
+import org.quartz.DateBuilder;
+import org.quartz.JobDetail;
+import org.quartz.Scheduler;
+import org.quartz.SchedulerException;
+import org.quartz.SchedulerFactory;
+import org.quartz.SimpleTrigger;
+import org.quartz.impl.StdSchedulerFactory;
 
-  String JAVA_TIMER_NAME = "JavaTimer";
+public class Timer {
+  private final Scheduler sched;
+  private final JobDetail job;
+  private final UUID uuid;
+  private int startDelay;
+  private int interval;
 
-  void start();
+  public Timer(EventBus eventBus, Integer startDelay, Integer interval)
+      throws IllegalArgumentException {
+    SchedulerFactory sf = new StdSchedulerFactory();
+    this.startDelay = startDelay;
+    this.interval = interval;
+    try {
+      sched = sf.getScheduler();
+      uuid = UUID.randomUUID();
+      job =
+          newJob(ScheduledEvent.class)
+              .withIdentity(EventBus.class.getSimpleName() + uuid.toString(), "group")
+              .build();
+      job.getJobDataMap().put(EventBus.class.getSimpleName(), eventBus);
 
-  void stop();
+    } catch (SchedulerException e) {
+      throw new IllegalArgumentException(
+          "In QuartzTimer a SchedulerException was thrown: " + e.toString());
+    }
+  }
 
-  static Class getTimerType(String className) {
-    switch (className) {
-      case Timer.QUARTZ_TIMER_NAME:
-        return QuartzTimer.class;
-      case Timer.JAVA_TIMER_NAME:
-        return JavaTimer.class;
-      default:
-        return QuartzTimer.class;
+  public void start() throws IllegalArgumentException {
+    try {
+      SimpleTrigger trigger =
+          newTrigger()
+              .withIdentity("trigger-" + EventBus.class.getSimpleName() + uuid.toString(), "group")
+              .startAt(DateBuilder.futureDate(startDelay, DateBuilder.IntervalUnit.MILLISECOND))
+              .withSchedule(simpleSchedule().withIntervalInMilliseconds(interval).repeatForever())
+              .build();
+      sched.scheduleJob(job, trigger);
+      sched.start();
+    } catch (SchedulerException e) {
+      throw new IllegalArgumentException(
+          "In QuartzTimer a SchedulerException was thrown: " + e.toString());
+    }
+  }
+
+  public void stop() {
+    try {
+      sched.shutdown();
+    } catch (SchedulerException e) {
+      throw new IllegalArgumentException(
+          "In QuartzTimer a SchedulerException was thrown: " + e.toString());
     }
   }
 }
